add helm support and more

+++ cp.sh
@@ -2,13 +2,22 @@
 
 BASE_DIR="$( cd -- "$(dirname "${0}")" >/dev/null 2>&1 ; pwd -P )"
 
-source "${BASE_DIR}/env.sh" "${BASE_DIR}" ;
-source "${BASE_DIR}/certs.sh" "${BASE_DIR}" ;
-source "${BASE_DIR}/helpers.sh" ;
-source "${BASE_DIR}/tsb-helpers.sh" ;
+# shellcheck source=/dev/null
+source "${BASE_DIR}/env.sh" ;
+# shellcheck source=/dev/null
+source "${BASE_DIR}/helpers/certs.sh" ;
+# shellcheck source=/dev/null
+source "${BASE_DIR}/helpers/print.sh" ;
+# shellcheck source=/dev/null
+source "${BASE_DIR}/helpers/registry.sh" ;
+# shellcheck source=/dev/null
+source "${BASE_DIR}/helpers/tsb.sh" ;
 
 ACTION=${1} ;
 
+TSB_HELM_REPO="https://charts.dl.tetrate.io/public/helm/charts" ;
+MP_HELM_CHART="tetrate-tsb-charts/managementplane" ;
+CP_HELM_CHART="tetrate-tsb-charts/controlplane" ;
 
 # This function provides help information for the script.
 #
@@ -19,45 +28,32 @@ function help() {
   echo "  --uninstall: uninstall tsb control and data plane" ;
 }
 
-# Patch OAP refresh rate of control plane
-#   args:
-#     (1) cluster context
-function patch_oap_refresh_rate_cp {
-  [[ -z "${1}" ]] && print_error "Please provide cluster context as 1st argument" && return 2 || local cluster_ctx="${1}" ;
-
-  local oap_patch='{"spec":{"components":{"oap":{"streamingLogEnabled":true,"kubeSpec":{"deployment":{"env":[{"name":"SW_CORE_PERSISTENT_PERIOD","value":"5"}]}}}}}}' ;
-  kubectl --context "${cluster_ctx}" -n istio-system patch controlplanes controlplane --type merge --patch "${oap_patch}" ;
-}
-
-
 # This function installs the tsb control and data plane using tctl.
 #
 function install_tctl() {
 
-  local certs_base_dir=$(get_certs_base_dir) ;
-  local install_repo_url=$(get_install_repo_url) ;
-  local mp_cluster_name=$(get_mp_name) ;
-  local mp_output_dir=$(get_mp_output_dir) ;
+  local certs_base_dir; certs_base_dir="$(get_certs_output_dir)" ;
+  local install_repo_url; install_repo_url=$(get_local_registry_endpoint) ;
+  local mp_cluster_name; mp_cluster_name=$(get_mp_name) ;
+  local mp_output_dir; mp_output_dir=$(get_mp_output_dir) ;
 
   # Login again as tsb admin in case of a session time-out
-  login_tsb_admin tetrate ;
+  login_tsb_admin "tetrate" "admin" "admin" ;
 
-  export TSB_API_ENDPOINT=$(kubectl --context "${mp_cluster_name}" get svc -n tsb envoy --output jsonpath='{.status.loadBalancer.ingress[0].ip}') ;
-  export TSB_INSTALL_REPO_URL=${install_repo_url} ;
 
   cp_count=$(get_cp_count) ;
   cp_index=0 ;
   while [[ ${cp_index} -lt ${cp_count} ]]; do
     cp_cluster_name=$(get_cp_name_by_index ${cp_index}) ;
-    cp_trust_domain=$(get_cp_trust_domain_by_index ${cp_index}) ;
-    cp_template_file=$(get_cp_template_file ${cp_index}) ;
+    cp_cr_template_file=$(get_cp_cr_template_file_by_index ${cp_index}) ;
     cp_output_dir=$(get_cp_output_dir ${cp_index}) ;
     print_info "Start installation of tsb control plane in cluster ${cp_cluster_name}" ;
 
     # Generate a service account private key for the active cluster
     #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#using-tctl-to-generate-secrets
-    kubectl config use-context "${mp_cluster_name}" ;
-    tctl install cluster-service-account --cluster "${cp_cluster_name}" > "${cp_output_dir}/cluster-service-account.jwk" ;
+    if [ ! -f "${cp_output_dir}/cluster-service-account.jwk" ]; then
+      tctl install cluster-service-account --cluster "${cp_cluster_name}" > "${cp_output_dir}/cluster-service-account.jwk" ;
+    fi
 
     # Create control plane secrets
     #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#using-tctl-to-generate-secrets
@@ -71,14 +67,14 @@ function install_tctl() {
       > "${cp_output_dir}/controlplane-secrets.yaml" ;
 
     # Generate controlplane.yaml by inserting the correct mgmt plane API endpoint IP address
-    export TSB_CLUSTER_NAME=${cp_cluster_name} ;
-    export TSB_TRUST_DOMAIN=${cp_trust_domain} ;
-    envsubst < "${cp_template_file}" > "${cp_output_dir}/controlplane.yaml" ;
+    TSB_API_SERVER_IP=$(get_tsb_api_ip "${mp_cluster_name}") ; export TSB_API_SERVER_IP ;
+    TSB_API_SERVER_PORT=$(get_tsb_api_port "${mp_cluster_name}") ; export TSB_API_SERVER_PORT ;
+    export TSB_INSTALL_REPO_URL=${install_repo_url} ;
+    envsubst < "${cp_cr_template_file}" > "${cp_output_dir}/controlplane.yaml" ;
 
     # bootstrap cluster with self signed certificate that share a common root certificate
     #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#intermediate-istio-ca-certificates
-    generate_istio_cert "${cp_cluster_name}" ;
-
+    generate_istio_cert "${certs_base_dir}" "${cp_cluster_name}" ;
     if ! kubectl --context "${cp_cluster_name}" get ns istio-system &>/dev/null; then
       kubectl --context "${cp_cluster_name}" create ns istio-system ;
     fi
@@ -92,7 +88,7 @@ function install_tctl() {
 
     # Deploy operators
     #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#deploy-operators
-    login_tsb_admin tetrate ;
+    login_tsb_admin "tetrate" "admin" "admin" ;
     tctl install manifest cluster-operators --registry ${install_repo_url} > ${cp_output_dir}/clusteroperators.yaml ;
 
     # Applying operator, secrets and control plane configuration
@@ -124,15 +120,13 @@ function install_tctl() {
     print_info "Finished installation of tsb control plane in cluster ${cp_cluster_name}" ;
     cp_index=$((cp_index+1)) ;
   done
-
-  exit 0 ;
 }
 
 # This function installs the tsb control and data plane using tctl.
 #
 function uninstall_tctl() {
 
-  local cp_count=$(get_cp_count) ;
+  local cp_count; cp_count=$(get_cp_count) ;
   local cp_index=0 ;
   while [[ ${cp_index} -lt ${cp_count} ]]; do
     cp_cluster_name=$(get_cp_name_by_index ${cp_index}) ;
@@ -196,11 +190,101 @@ function uninstall_tctl() {
     print_info "Finished removing installation of tsb control plane in cluster ${cp_cluster_name}" ;
     cp_index=$((cp_index+1)) ;
   done
-
-  exit 0 ;
 }
 
+# This function installs the tsb control plane using helm.
+#
+function install_helm {
 
+  local certs_base_dir; certs_base_dir="$(get_certs_output_dir)" ;
+  local install_repo_url; install_repo_url=$(get_local_registry_endpoint) ;
+  local mp_cluster_name; mp_cluster_name=$(get_mp_name) ;
+  local mp_output_dir; mp_output_dir=$(get_mp_output_dir) ;
+  local tsb_version; tsb_version=$(get_tsb_version) ;
+
+  if ! helm repo list | grep -q 'tetrate-tsb-charts'; then
+    helm repo add tetrate-tsb-charts "${TSB_HELM_REPO}" ;
+  fi
+  helm repo update tetrate-tsb-charts ;
+
+  cp_count=$(get_cp_count) ;
+  cp_index=0 ;
+  while [[ ${cp_index} -lt ${cp_count} ]]; do
+    cp_cluster_name=$(get_cp_name_by_index ${cp_index}) ;
+    cp_helm_template_file=$(get_cp_helm_template_file_by_index ${cp_index}) ;
+    cp_output_dir=$(get_cp_output_dir ${cp_index}) ;
+    print_info "Start installation of tsb control plane in cluster ${cp_cluster_name}" ;
+
+    # Generate a service account private key for the mgmt cp cluster
+    #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#using-tctl-to-generate-secrets
+    tctl install cluster-service-account --cluster "${cp_cluster_name}" > "${cp_output_dir}/cluster-service-account.jwk" ;
+
+    TSB_API_SERVER_IP=$(get_tsb_api_ip "${mp_cluster_name}") ; export TSB_API_SERVER_IP ;
+    TSB_API_SERVER_PORT=$(get_tsb_api_port "${mp_cluster_name}") ; export TSB_API_SERVER_PORT ;
+    TSB_INSTALL_REPO_URL="${install_repo_url}" ; export TSB_INSTALL_REPO_URL ;
+    export TSB_VERSION="${tsb_version}" ;
+    envsubst < "${cp_helm_template_file}" > "${cp_output_dir}/cp-helm-values.yaml" ;
+
+    # bootstrap cluster with self signed certificate that share a common root certificate
+    #   REF: https://docs.tetrate.io/service-bridge/1.6.x/en-us/setup/self_managed/onboarding-clusters#intermediate-istio-ca-certificates
+    generate_istio_cert "${certs_base_dir}" "${cp_cluster_name}" ;
+    if ! kubectl --context "${cp_cluster_name}" get ns istio-system &>/dev/null; then
+      kubectl --context "${cp_cluster_name}" create ns istio-system ;
+    fi
+    if ! kubectl --context "${cp_cluster_name}" -n istio-system get secret cacerts &>/dev/null; then
+      kubectl --context "${cp_cluster_name}" create secret generic cacerts -n istio-system \
+        --from-file="${certs_base_dir}/${cp_cluster_name}/ca-cert.pem" \
+        --from-file="${certs_base_dir}/${cp_cluster_name}/ca-key.pem" \
+        --from-file="${certs_base_dir}/${cp_cluster_name}/root-cert.pem" \
+        --from-file="${certs_base_dir}/${cp_cluster_name}/cert-chain.pem" ;
+    fi
+
+    # install tsb control plane using helm
+    helm upgrade --install tsb-cp "${CP_HELM_CHART}" \
+      --create-namespace \
+      --kube-context "${cp_cluster_name}" \
+      --namespace "istio-system" \
+      --set-file secrets.elasticsearch.cacert="${mp_output_dir}/tsb-ca-cert.pem" \
+      --set-file secrets.tsb.cacert="${mp_output_dir}/tsb-ca-cert.pem" \
+      --set-file secrets.xcp.rootca="${mp_output_dir}/tsb-ca-cert.pem" \
+      --set-file secrets.clusterServiceAccount.JWK="${cp_output_dir}/cluster-service-account.jwk" \
+      --values "${cp_output_dir}/cp-helm-values.yaml" \
+      --wait ;
+    cp_index=$((cp_index+1)) ;
+  done
+
+  cp_count=$(get_cp_count) ;
+  cp_index=0 ;
+  while [[ ${cp_index} -lt ${cp_count} ]]; do
+    cp_cluster_name=$(get_cp_name_by_index ${cp_index}) ;
+    print_info "Waiting for tsb control plane deployments to become available" ;
+    wait_cp_ready "${cp_cluster_name}" "istio-system" ;
+
+    kubectl --context "${cp_cluster_name}" get pods -A ;
+    print_info "Finished installation of tsb control plane in cluster ${cp_cluster_name}" ;
+    cp_index=$((cp_index+1)) ;
+  done
+}
+
+# This function uninstalls the tsb control plane using helm.
+#
+function uninstall_helm {
+  local mp_cluster_name; mp_cluster_name=$(get_mp_name) ;
+
+  cp_count=$(get_cp_count) ;
+  cp_index=0 ;
+  while [[ ${cp_index} -lt ${cp_count} ]]; do
+    cp_cluster_name=$(get_cp_name_by_index ${cp_index}) ;
+    echo "Uninstalling helm chart for tsb management/control plane in cluster ${cp_cluster_name}" ;
+    helm uninstall tsb-cp \
+      --kube-context "${cp_cluster_name}" \
+      --namespace "istio-system" ;
+    kubectl --context "${cp_cluster_name}" delete namespace "istio-system" ;
+    cp_index=$((cp_index+1)) ;
+  done
+
+  print_info "Uninstalled helm chart for tsb management/control plane" ;
+}
 
 # Main execution
 #
@@ -209,12 +293,28 @@ case "${ACTION}" in
     help ;
     ;;
   --install)
-    print_stage "Going to install tsb control and data plane" ;
-    install_tctl ;
+    if [[ "$(get_tsb_install_method)" == "helm" ]]; then
+      print_stage "Going to install tsb control and data plane using helm" ;
+      install_helm ;
+    elif [[ "$(get_tsb_install_method)" == "tctl" ]]; then
+      print_stage "Going to install tsb ontrol and data plane using tctl" ;
+      install_tctl ;
+    else
+      print_error "Invalid tsb install method. Choose 'helm' or 'tctl'" ;
+      help ;
+    fi
     ;;
   --uninstall)
-    print_stage "Going to uninstall tsb control and data plane" ;
-    uninstall_tctl ;
+    if [[ "$(get_tsb_install_method)" == "helm" ]]; then
+      print_stage "Going to uninstall tsb ontrol and data plane using helm" ;
+      uninstall_helm ;
+    elif [[ "$(get_tsb_install_method)" == "tctl" ]]; then
+      print_stage "Going to uninstall tsb ontrol and data plane using tctl" ;
+      uninstall_tctl ;
+    else
+      print_error "Invalid tsb install method. Choose 'helm' or 'tctl'" ;
+      help ;
+    fi
     ;;
   *)
     print_error "Invalid option. Use 'help' to see available commands." ;
